Consolidate project page code.
