+++ frontend/src/pages/Project.svelte
@@ -2,13 +2,15 @@
     import Navbar from "../components/Navbar.svelte";
     import PageHeader from "../components/PageHeader.svelte";
     import ProjectVM from "../components/ProjectVM.svelte";
-    import {Projects, Snackbars, User} from "../stores";
+    import { Projects, User } from "../stores";
     import PageTitle from "../components/PageTitle.svelte";
     import Input from "../components/Input.svelte";
     import Button from "../components/Button.svelte";
-    import {push} from "svelte-spa-router";
-    import Proxies from "../components/Proxies.svelte";
+    import { push } from "svelte-spa-router";
+    import Proxies from "../components/project/Proxies.svelte";
     import AuditLog from "./AuditLog.svelte";
+    import { checkMeta, updateProjects } from "../utils";
+import Settings from "../components/project/Settings.svelte";
 
     export let params: any = {};
 
@@ -23,78 +25,11 @@
         return returnProject;
     };
 
-    let views = ["RESOURCES", "SETTINGS", "PROXIES", "AUDIT LOG"];
-    let currentView = "RESOURCES";
-
+    let project: Project;
     $: project = getProjectById(params.project_id, $Projects);
 
-    function toVM(vm: any): VM {
-        return vm as VM;
-    }
-
-    let newUserEmail;
-    let newBudget = 0;
-
-    function submitAddUser() {
-        fetch("__apiRoute__/project/adduser", {
-            method: "POST",
-            headers: {"Content-Type": "application/json"},
-            body: JSON.stringify({project: params.project_id, email: newUserEmail})
-        })
-            .then(resp => resp.json())
-            .then(data => {
-                alert(data.meta.message);
-                location.reload(); // FIXME: This probably not the best solution, you should make it call correct functions from utils.
-            })
-            .catch((err) => alert(err));
-    }
-
-    function changeBudget() {
-        fetch("__apiRoute__/project/changebudget", {
-            method: "POST",
-            headers: {"Content-Type": "application/json"},
-            body: JSON.stringify({project: params.project_id, budget: newBudget})
-        })
-            .then(resp => resp.json())
-            .then(data => {
-                alert(data.meta.message);
-                location.reload(); // FIXME: This probably not the best solution, you should make it call correct functions from utils.
-            })
-            .catch((err) => alert(err));
-    }
-
-    function deleteProject() {
-        if (confirm("Are you sure you want to delete this project? This action cannot be undone.")) {
-            fetch("__apiRoute__/project", {
-                method: "DELETE",
-                headers: {"Content-Type": "application/json"},
-                body: JSON.stringify({project: params.project_id})
-            })
-                .then(resp => resp.json())
-                .then(data => {
-                    if (data.meta.success) {
-                        $Snackbars.push({
-                            color: "green",
-                            status: 200,
-                            message: data.meta.message,
-                            grouped: true,
-                        })
-                        $Snackbars = $Snackbars;
-                        push("/dashboard");
-                    } else {
-                        $Snackbars.push({
-                            color: "red",
-                            status: "ERROR",
-                            message: data.meta.message,
-                            grouped: true,
-                        })
-                        $Snackbars = $Snackbars;
-                    }
-                })
-                .catch((err) => alert(err));
-        }
-    }
-
+    let views = ["RESOURCES", "SETTINGS", "PROXIES", "AUDIT LOG"];
+    let currentView = "RESOURCES";
 </script>
 
 <PageTitle title={project ? project.name : 'Project page'}/>
@@ -105,7 +40,7 @@
 			{label: 'Dashboard', path: '/dashboard'},
 			{label: project.name, path: `/dashboard/projects/${project._id}`},
 			]}/>
-        <PageHeader options bind:current={currentView} labels={views} baseHref={"/#/dashboard/projects/" + params.project_id} {params}>
+        <PageHeader options labels={views} baseHref={"/#/dashboard/projects/" + params.project_id} {params}>
             {project.name}
         </PageHeader>
         <div class="content">
@@ -113,8 +48,10 @@
                 <span class="title">Virtual Machines</span><br>
                 <span class="subtitle">{project.budget_used} out of {project.budget} allocated {project.budget === 1 ? "cores" : "core"} used</span>
                 {#if project.vms.length === 0}
-                    <div class="empty-list">Nothing to see here...</div>
-                    <a href="/#/dashboard/create" class="add-new-button"> CREATE VM </a>
+                    <div class="empty-list">
+                        Nothing to see here...<br>
+                        <Button href="/#/dashboard/create" style="margin-top: 20px;">CREATE VM</Button>
+                    </div>
                 {:else}
 					<span class="labels">
 						<div class="hostname-label">HOSTNAME</div>
@@ -123,38 +60,12 @@
 					</span>
                     <div class="vm-list">
                         {#each project.vms as vm}
-                            <ProjectVM VM={toVM(vm)} link={'/dashboard/projects/' + project._id + '/resources/' + vm['_id']}/>
+                            <ProjectVM VM={vm} link={'/dashboard/projects/' + project._id + '/resources/' + vm['_id']}/>
                         {/each}
                     </div>
                 {/if}
             {:else if params.page === "settings"}
-                <div>
-                    <span class="title">Settings</span>
-                    <div class="user-form-container">
-                        <span class="user-form-subheader">Users:</span>
-                        <ul>
-                            {#each project.users as user}
-                                <li>{user}</li>
-                            {/each}
-                        </ul>
-                    </div>
-
-                    <div class="user-form-container">
-                        <span class="user-form-subheader">Add user to project:</span>
-                        <span class="user-form-subtitle">Enter a user's email. Make sure they already have signed up for an account.</span>
-                        <Input bind:value={newUserEmail} autocomplete="off" type="text" class="user-input" placeholder="user@example.com"/>
-                        <Button width="150px" on:click={() => submitAddUser()}>ADD USER</Button>
-                        {#if $User.admin}
-                            <span style="margin-top: 2rem" class="user-form-subheader">Set project budget:</span>
-                            <span class="user-form-subtitle">Current usage is {project.budget_used}/{project.budget}</span>
-                            <Input bind:value={newBudget} autocomplete="off" type="number" class="user-input" placeholder="2"/>
-                            <Button width="220px" on:click={() => changeBudget()}>CHANGE BUDGET</Button>
-                        {/if}
-                        <!-- <button class="user-form-button" on:click={() => submitAddUser()}>ADD USER</button> -->
-                        <!-- TODO: THIS SHOULD ONLY SHOW FOR PROJECT CREATOR (OR OWNER?) -->
-                        <Button width="250px" color="#aa1717" style="margin-top: 3rem;" on:click={deleteProject}>DELETE PROJECT</Button>
-                    </div>
-                </div>
+                <Settings {project} {params}/>
             {:else if params.page === "proxies"}
                 <Proxies {project}/>
             {:else if params.page === "auditlog"}
@@ -172,32 +83,11 @@
         flex-direction: column;
     }
 
-    .add-new-button {
-        height: 40px;
-        display: flex;
-        align-items: center;
-        justify-content: center;
-        font-weight: 500;
-        color: white;
-        background-color: #0e0d0d;
-        border: none;
-        font-family: inherit;
-        margin: 0 auto;
-        font-size: 16px;
-        padding: 0 30px;
-        text-decoration: none;
-        width: 150px;
-    }
-
-    button:active {
-        opacity: 0.8;
-    }
-
     div.empty-list {
         font-weight: 500;
         text-align: center;
         padding-bottom: 10px;
-        opacity: 0.8;
+        color: #575656;
     }
 
     .content {
@@ -252,28 +142,4 @@
         display: flex;
         justify-content: center;
     }
-
-    .user-form-container {
-        display: flex;
-        flex-direction: column;
-        margin-top: 25px;
-        padding-left: 15px;
-    }
-
-    .user-form-subheader {
-        font-size: 22px;
-        font-weight: 500;
-    }
-
-    .user-form-subtitle {
-        font-size: 16px;
-        opacity: 0.5;
-        padding-bottom: 10px;
-    }
-
-    div :global(.user-input) {
-        margin: 0 0 10px 0;
-        width: 350px;
-    }
-
 </style>
